Create BuscadorLlamadas.py
Many tables and column names were renamed due to my company's data privacy policies. However, there are no restrictions on sharing the code I created for the company. A ".env" file was used to handle credentials, so you will need to create it and place it in the same environment to work with the same code.

Once exported to .exe, the example file weighed approximately 60 MB due to the encapsulation of Python libraries and the files used within the graphical interface (button images). As a recommendation, I suggest creating the buttons directly in Tkinter. Due to time constraints, I decided to design them in Photoshop, but it is not the most efficient approach.